requirements.txt
alembic==1.14.0
APScheduler==3.11.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Mako==1.3.8
MarkupSafe==3.0.2
paramiko==3.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
sshtunnel==0.4.0
typing_extensions==4.12.2
tzlocal==5.2
